feat(pim): todoman
i prefer taskwarrior as my source of truth, but todoman allows me to
more seamlessly interoperate with (and backup) my iphone reminders
without activating iCloud and other vendor lockin traps.

in other words: the main feature of taskwarrior is that it's where work
gets done. the main feature of todoman is that it's standards-compliant,
relying on iCal / VTODO, and can be used anywhere. my linux laptop can
now be a bridge between the two.